Refactor mapping into Adapter class in order to make overwriting data and bus mapping possible within adapter
